backend/requirements.txt
asgiref==3.7.2
cssselect==1.1.0
django==4.2.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.5.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
git+https://https://github.com/VidipKhattar/pytube


